pyproject.toml
[project]
name = "prml-llm-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.4.2",
    "dotenv>=0.9.9",
    "loguru>=0.7.3",
    "pre-commit>=4.5.1",
    "pyyaml>=6.0.1",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
]
